+++ include/RMF/decorators.h
@@ -18,6 +18,8 @@
 #include "RMF/decorator/publication.h"
 #include "RMF/decorator/sequence.h"
 #include "RMF/decorator/shape.h"
+#include "RMF/decorator/bond.h"
+#include "RMF/decorator/representation.h"
 #include "RMF/decorator/alternatives.h"
 
 /** \namespace RMF::decorator