+++ src/decorator/alternatives.cpp
@@ -15,17 +15,13 @@ namespace RMF {
 namespace decorator {
 AlternativesFactory::AlternativesFactory(FileHandle fh)
     : cat_(fh.get_category("alternatives")),
-      base_resolution_key_(fh.get_key<FloatTraits>(cat_, "resolution")),
       types_key_(fh.get_key<IntsTraits>(cat_, "types")),
-      roots_key_(fh.get_key<IntsTraits>(cat_, "roots")),
-      resolutions_key_(fh.get_key<FloatsTraits>(cat_, "resolutions")) {}
+      roots_key_(fh.get_key<IntsTraits>(cat_, "roots")) {}
 
 AlternativesFactory::AlternativesFactory(FileConstHandle fh)
     : cat_(fh.get_category("alternatives")),
-      base_resolution_key_(fh.get_key<FloatTraits>(cat_, "resolution")),
       types_key_(fh.get_key<IntsTraits>(cat_, "types")),
-      roots_key_(fh.get_key<IntsTraits>(cat_, "roots")),
-      resolutions_key_(fh.get_key<FloatsTraits>(cat_, "resolutions")) {}
+      roots_key_(fh.get_key<IntsTraits>(cat_, "roots")) {}
 
 namespace {
 double get_resolution_metric(double a, double b) {
@@ -33,24 +29,25 @@ double get_resolution_metric(double a, double b) {
   return a / b - 1;
 }
 
-NodeID get_alternative_impl(NodeConstHandle cur, FloatKey base_resolution_key,
-                            IntsKey types_key, IntsKey roots_key,
-                            FloatsKey resolutions_key, RepresentationType type,
+NodeID get_alternative_impl(NodeConstHandle cur, IntsKey types_key,
+                            IntsKey roots_key, RepresentationType type,
                             float resolution) {
   RMF_USAGE_CHECK(type == PARTICLE, "Only particle type supported currently");
   if (!cur.get_has_value(types_key)) return cur.get_id();
 
-  double closest_resolution = cur.get_value(base_resolution_key);
+  double closest_resolution = get_resolution(cur);
   int closest_index = -1;
   Nullable<IntsTraits> types = cur.get_value(types_key);
   if (!types.get_is_null()) {
-    Nullable<FloatsTraits> resolutions = cur.get_value(resolutions_key);
+    Ints roots = cur.get_value(roots_key);
     for (unsigned int i = 0; i < types.get().size(); ++i) {
-      if (get_resolution_metric(resolution, resolutions.get()[i]) <
-              get_resolution_metric(resolution, closest_resolution) &&
-          types.get()[i] == type) {
+      if (types.get()[i] != type) continue;
+      double cur_resolution =
+          get_resolution(cur.get_file().get_node(NodeID(roots[i])));
+      if (get_resolution_metric(resolution, cur_resolution) <
+          get_resolution_metric(resolution, closest_resolution)) {
         closest_index = i;
-        closest_resolution = resolutions.get()[i];
+        closest_resolution = cur_resolution;
       }
     }
   }
@@ -87,28 +84,18 @@ std::pair<double, double> get_resolution_impl(
   return ret;
 }
 
-float get_resolution(std::pair<double, double> total) {
-  return total.second / total.first;
 }
 
-float get_resolution_impl(NodeConstHandle root) {
+double get_resolution(NodeConstHandle root) {
   IntermediateParticleFactory ipcf(root.get_file());
   std::pair<double, double> total = get_resolution_impl(root, ipcf);
   RMF_USAGE_CHECK(total.first != 0,
                   std::string("No particles were found at ") + root.get_name());
-  return get_resolution(total);
-}
+  return total.second / total.first;
 }
 
-Alternatives::Alternatives(NodeHandle nh, FloatKey base_resolution_key,
-                           IntsKey types_key, IntsKey roots_key,
-                           FloatsKey resolutions_key)
-    : AlternativesConst(nh, base_resolution_key, types_key, roots_key,
-                        resolutions_key) {
-  if (!nh.get_has_value(base_resolution_key_)) {
-    nh.set_value(base_resolution_key_, get_resolution_impl(nh));
-  }
-}
+Alternatives::Alternatives(NodeHandle nh, IntsKey types_key, IntsKey roots_key)
+    : AlternativesConst(nh, types_key, roots_key) {}
 
 void Alternatives::add_alternative(NodeHandle root, RepresentationType type) {
   get_node()
@@ -117,50 +104,32 @@ void Alternatives::add_alternative(NodeHandle root, RepresentationType type) {
       .push_back(type);
   get_node()
       .get_shared_data()
-      ->access_static_value(get_node().get_id(), resolutions_key_)
-      .push_back(get_resolution_impl(root));
-  get_node()
-      .get_shared_data()
       ->access_static_value(get_node().get_id(), roots_key_)
       .push_back(root.get_id().get_index());
 }
 
 NodeConstHandle AlternativesConst::get_alternative(RepresentationType type,
                                                    double resolution) const {
-  return get_node().get_file().get_node(
-      get_alternative_impl(get_node(), base_resolution_key_, types_key_,
-                           roots_key_, resolutions_key_, type, resolution));
+  return get_node().get_file().get_node(get_alternative_impl(
+      get_node(), types_key_, roots_key_, type, resolution));
 }
 
 NodeConstHandles AlternativesConst::get_alternatives(RepresentationType type)
     const {
   RMF_USAGE_CHECK(type == PARTICLE, "Only particles supported");
   NodeConstHandles ret;
-  RMF_FOREACH(NodeID id,
+  RMF_FOREACH(NodeID nid,
               get_alternatives_impl(get_node(), types_key_)) {
-    ret.push_back(get_node().get_file().get_node(id));
+    ret.push_back(get_node().get_file().get_node(nid));
   }
   return ret;
-}
-
-float AlternativesConst::get_resolution(NodeID id) const {
-  if (id == get_node().get_id()) {
-    return get_node().get_value(base_resolution_key_);
-  }
-  const Ints& roots = get_node().get_value(roots_key_);
-  for (unsigned int i = 0; i < roots.size(); ++i) {
-    if (roots[i] == static_cast<int>(id.get_index())) {
-      return get_node().get_value(resolutions_key_).get()[i];
-    }
-  }
-  RMF_THROW(Message("No such alternative representation"), UsageException);
 }
 
 RepresentationType AlternativesConst::get_representation_type(NodeID id) const {
   if (id == get_node().get_id()) {
     return PARTICLE;
   }
-  const Ints& roots = get_node().get_value(roots_key_);
+  Ints roots = get_node().get_value(roots_key_);
   for (unsigned int i = 0; i < roots.size(); ++i) {
     if (roots[i] == static_cast<int>(id.get_index())) {
       return RepresentationType(get_node().get_value(types_key_).get()[i]);
@@ -175,7 +144,7 @@ Floats get_resolutions_impl(NodeConstHandle root, AlternativesFactory af,
   Floats ret;
   if (af.get_is(root)) {
     RMF_FOREACH(NodeConstHandle a, af.get(root).get_alternatives(type)) {
-      ret.push_back(af.get(root).get_resolution(a));
+      ret.push_back(get_resolution(a));
     }
   } else {
     RMF_FOREACH(NodeConstHandle ch, root.get_children()) {